pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "charlie-agents"
version = "0.1.0"
description = "Universal command transpiler for AI agents, MCP servers, and rules"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Charlie Contributors"}
]
keywords = ["ai", "agents", "mcp", "commands", "transpiler", "claude", "copilot", "gemini"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "typer>=0.9.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
charlie = "charlie.cli:app"

[project.urls]
Homepage = "https://github.com/henriquemoody/charlie"
Documentation = "https://github.com/henriquemoody/charlie#readme"
Repository = "https://github.com/henriquemoody/charlie"
Issues = "https://github.com/henriquemoody/charlie/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/charlie"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=charlie --cov-report=term-missing --cov-report=xml"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = "tests.*"
# Allow tests to be more flexible with object construction and fixtures
disallow_untyped_defs = false
disable_error_code = ["call-arg", "arg-type", "union-attr", "no-any-return", "func-returns-value", "override", "assignment"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []

